+++ src/core/src/storage_class.cpp
@@ -5,7 +5,6 @@
 namespace melon::core
 {
 
-
 User::User(std::uint64_t user_id, std::string username, Status status)
     : m_user_id(user_id)
     , m_username(std::move(username))
@@ -13,9 +12,8 @@ User::User(std::uint64_t user_id, std::string username, Status status)
 {
 }
 
-
 Message::Message(std::uint64_t message_id, std::uint64_t user_id,
-                 std::uint32_t chat_id, std::string text, Message::Status status)
+                 std::uint64_t chat_id, std::string text, Message::Status status)
     : m_message_id(message_id)
     , m_user_id(user_id)
     , m_chat_id(chat_id)
@@ -24,8 +22,7 @@ Message::Message(std::uint64_t message_id, std::uint64_t user_id,
 {
 }
 
-
-Chat::Chat(std::uint32_t chat_id, std::string chatname)
+Chat::Chat(std::uint64_t chat_id, std::string chatname)
     : m_chat_id(chat_id)
     , m_chatname(std::move(chatname))
 {
